pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "transcription-service"
version = "0.1.0"
description = "Servicio que gestiona la ganereación de texto a partir de audio/video"
authors = ["GEB Team <team@geb.com>"]
readme = "README.md"
packages = [{include = "src"}]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
fastapi = "^0.122.0"
uvicorn = "^0.38.0"
requests = "^2.32.5"
pydantic-settings = "^2.12.0"
azure-storage-blob = "^12.27.1"
numpy = "^2.3.5"
azure-servicebus = "^7.14.3"
tenacity = "^8.2"
azure-identity = ">=1.25.0"
kink = "^0.8.1"
openai = "^2.9.0"
pymongo = "^4.15.5"
motor = ">=3.6.0"
aiohttp = ">=3.13.3"
openpyxl = ">=3.1.5"

[tool.poetry.group.test.dependencies]
pytest = ">=7.4"
pytest-asyncio = ">=0.23"
coverage = ">=7.10.7"
pytest-cov = ">=7.0.0"
pysonar = ">=1.2.0.2419"

[tool.poetry.scripts]
transcription-service = "src.main:run"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
